Remove "do-nothing" destructors.
No reason to have them around, C++ gives us the same thing for free so why not take advantage of that when we can? Not to mention that I was inconsistently supplying them for some objects but not others; better to be consistent, and it is easier to do so when I go the "easy" route and take advantage of the defaults anyway.