Tile objects can now hold and display an Actor.
This may be extended in the future to allow a Tile to hold multiple Actors, but for now only a single Actor can held in a Tile.

There is currently no checking for nor handling of situations where an Actor is being added to a Tile that may already have an Actor in it.