Tried out a layering approach, unsuccessfully.
I attempted to add an "overlay" over top of base level window with the idea that it could be used to display entities independently of the level/tiles themselves. Unfortunately, ncurses does not actually work like this, and subwindows are nothing more than different coordinate systems that access identical memory locations and, thus, we can only see what is displayed last in the code.