addComponent method now expects a pointer.
This allows me to do pointer casts when receiving the Component pointer
and treat it as the proper Component sub-class by simply re-casting the
pointer. Now that this seems to be working, however, I'm not really
happy with the implications of how this would be used, and I'm leaning
now toward ripping out the entire concept of Components existing as
anything more than internal collections of raw data, and instead
providing individual managers with specialized methods to get at said
data.