SkillManager will now account for attribute damage
When calculating skill level, the level is penalized by 1 point for each
full 20% that the controlling attribute has been damaged by. This also
necessitates adding another constant in SkillComponent.h to map skills
to their controlling attributes.

Unfortunately, there's a problem somewhere with the math, and it doesn't
work as expected right now.