Stubbed out managers for world data
WorldManager will eventually keep track of the whole world, including the various dungeons within it; each dungeon will have an associated DungeonManager that will track its data, including each of the levels within it; each of those level will be tracked by a LevelManager. Eventually all of these will take parameters to control and refine how they are generated; for now it is just one unnecessary heap of fugly because I have to reach around it awkwardly to get to the data I need.