Coerce libtcod to find DLLs in lib dir
A simple hack, we cd into the lib dir before loading the Windows DLL, since said DLL looks in the current working directory to itself then load SDL.dll. We cd back to where we started before we continue on, just to make sure everything else works as expected (and it does!). This only happens on Windows; there is not an SDL library to link with on Linux (hidden external dependency?).