Rewrite how events are canceled
Rather than expecting all event handlers to return one of three different constants, the propagation of an event is now canceled by calling the stop() method on that event. This results in cleaner, simpler event handlers, makes it more explicit when an event is being stopped, and solves the bug of an event being stopped by a handler that merely omits any return statement -- all with one fell swoop!