Tile unlit color now just blend
Rather than trying to come up with an unlit Tile color ourselves (or relying on a jarringly-contrasting default value), we now define the unlit color to be the HSV blend of the lit color and the background (i.e. black); for now at least we use a straight 50% blend, i.e. the midpoint between the two colors.