Ditch PenetratingDamageEvent
I was not thinking in proper event-driven manner; the way damage/penetrating damage should be handled is to simply dispatch a single DamageEvent and allow another system -- e.g. an EquipmentSystem -- to modify it by applying armor effects before it is finally processed and damage is applied to the target.